# trace:  Use non-generic to replace newEvictedQueue in trace.start to reduce memory usage.